feat(DTFS2-7306): rename merge-tfm-values-into-portal-deals
